django/requirements.txt
Django==2.1.7
pytz==2018.9
psycopg2-binary
django-money
djangorestframework
djangorestframework-gis
coreapi
pyyaml
